added main window

+++ DB Scripts/create users table.sql
@@ -1,5 +1,5 @@
 create table users (
-    id  int not null,
+    id  number(5) not null,
     firstName char(30) not null,
     lastName char(30) not null,
     email char(50) unique,
